webcraw'g

+++ crawler/getlinks.py
@@ -3,6 +3,8 @@ from bs4 import BeautifulSoup
 from totext import seperatingFiles
 import logging
 import time
+from os import path
+from linksep import separatelinks
 
 logging.basicConfig(format="%(asctime)s - %(message)s", level=logging.INFO)
 
@@ -33,39 +35,46 @@ def cleaning_links(links: list):
 
 
 if __name__ == "__main__":
-    # getting all the links from the main page
-    total = get_href(links_home("https://www.rongovarsity.ac.ke/"))
-    # clean to remove ll ID links
-    total = set(cleaning_links(total))
-    other_links = []
-    # loop in each link generated from main page(total) to prodeuce additional links then save then inside other_links variable
-    for link in total:
-        time.sleep(5)
-        try:
-            other_links.append(', '.join(cleaning_links(get_href(links_home(link)))))
-            print(other_links)
-        except:
-            print(f'their is problem with {link}')
-    print(other_links)
-total_links = total.union(set(other_links))
-# saving the generated links inside txt file
-with open('../links.txt', 'w') as f:
-    for line in total_links:
-        f.write(line)
-        f.write('\n')
+    if not path.exists("../links.txt"):
+        # getting all the links from the main page
+        total = get_href(links_home("https://www.rongovarsity.ac.ke/"))
+        # clean to remove ll ID links
+        total = set(cleaning_links(total))
+        other_links = []
+        # loop in each link generated from main page(total) to prodeuce additional links then save then inside other_links variable
+        for link in total:
+            time.sleep(5)
+            try:
+                other_links.append(', '.join(cleaning_links(get_href(links_home(link)))))
+                print(other_links)
+            except:
+                print(f'their is problem with {link}')
+        print(other_links)
+        total_links = total.union(set(other_links))
+        # saving the generated links inside txt file
+        with open('../links.txt', 'w') as f:
+            for line in total_links:
+                f.write(line)
+                f.write('\n')
+    else:
 
 
-# getting data from website and saving it into .txt file
-def webpage_to_txt(links: set):
-    count = 0
-    for url in links:
-        logging.info(f"Getting data from {count} out of {len(links)}")
-        try:
-            separatingFiles(url)
-        except:
-            logging.info(f"Error Getting data from {url}")
-            continue
-        count += 1
+        # getting data from links extracted and saving it into .txt file
+        separatelinks()
+        def webpage_to_txt(links: str):
+            count = 0
+            for url in links:
+                print(url)
+                # logging.info(f"Getting data from {count} out of {len(links)}")
+                try:
+                    separatingFiles(url)
+                except:
+                    logging.info(f"Error Getting data from {url}")
+                    continue
+                count += 1
 
-
-webpage_to_txt(total_links)
+        with open('../links.txt', 'r') as f:
+            data = f.read()
+            print(data)
+            webpage_to_txt(data)
+            f.close()
